backend/requirements/prod.txt
Django==5.1.4
djangorestframework==3.15.2
celery==5.4.0
redis==5.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
requests==2.32.3
neo4j==5.28.1
python-json-logger==2.0.7
whitenoise==6.8.2
django-prometheus==2.3.1
beautifulsoup4==4.12.3
feedparser==6.0.11
lxml==5.3.0
minio==7.2.8
opensearch-py==2.7.1
python-dateutil==2.9.0.post0
deep-translator==1.11.4
openai==1.82.0
